+++ Solgema/fullcalendar/browser/adapters.py
@@ -8,6 +8,7 @@ from AccessControl import getSecurityManager
 from zope.interface import implements, Interface
 from plone.event.interfaces import IEventAccessor
 from zope.component import queryAdapter, adapts, getMultiAdapter, getAdapters
+from types import GeneratorType
 try:
     from Products.ZCatalog.interfaces import ICatalogBrain
 except:
@@ -56,6 +57,7 @@ try:
 except:
     class IDXCollection(Interface): pass
     class IFolder(Interface): pass
+from plone.app.querystring import queryparser
 
 
 def handle_recurrence(request):
@@ -144,7 +146,7 @@ def dict_from_events(events,
         else:
             raise ValueError('item type not supported for: %s' % repr(item))
 
-    if not hasattr(events, '__iter__'):
+    if not isinstance(events, (list, tuple, GeneratorType)):
         events = [events]
 
     return [dict_from_item(item) for item in events]
@@ -330,7 +332,7 @@ class SolgemaFullcalendarEventDict(object):
         referer = self.request.get('HTTP_REFERER')
         if referer:
             portal = getToolByName(self.context, 'portal_url').getPortalObject()
-            url = '/' + portal.id + referer.replace(portal.absolute_url(), '')
+            url = "/".join(portal.getPhysicalPath()) + referer.replace(portal.absolute_url(), '')
             context = portal.restrictedTraverse(url)
         eventPhysicalPath = '/'.join(event.getPhysicalPath())
         wft = getToolByName(context, 'portal_workflow')
@@ -1011,6 +1013,21 @@ class CollectionEventSource(TopicEventSource):
         return result
 
 
+def convert(value):
+    query = value['query']
+    if isinstance(query, unicode):
+        query = query.encode("utf-8")
+    elif isinstance(query, list):
+        query = [
+            item.encode("utf-8") if isinstance(item, unicode) else item
+            for item in query
+        ]
+    else:
+        pass
+    value['query'] = query
+    return value
+
+
 class DXCollectionEventSource(TopicEventSource):
     """Event source that get events from the collection
     """
@@ -1023,12 +1040,10 @@ class DXCollectionEventSource(TopicEventSource):
 
         listCriteria = context.query
 
-        # Handle operator-only query strings accordingly.
-        # Also convert to str since unicode breaks the catalog-query
-        query = dict(['v' in a and (str(a['i']),
-            [str(value) for value in a['v']]) or
-            (str(a['i']), [str(value) for value in a['o']])
-            for a in listCriteria])
+        query = dict([
+            (key, convert(value))
+            for key, value in queryparser.parseFormquery(context, listCriteria).items()
+        ])
 
         topicCriteria = interfaces.IListCriterias(context)()
         _args = {}